+++ tests/Pest.php
@@ -2,6 +2,9 @@
 
 declare(strict_types=1);
 
+use LaravelAnalytics\LaravelAnalytics\Tests\DatabaseTestCase;
 use LaravelAnalytics\LaravelAnalytics\Tests\TestCase;
 
-uses(TestCase::class)->in(__DIR__);
+uses(TestCase::class)->in('Feature', 'Unit');
+
+uses(DatabaseTestCase::class)->in('Database');